New Replays: Support pagination in search results
