+++ website/replays/index.template.html
@@ -91,6 +91,20 @@
 	.button {
 		vertical-align: middle;
 	}
+	.replay-controls {
+		padding-top: 10px;
+	}
+	.replay-controls h1 {
+		font-size: 16pt;
+		font-weight: normal;
+		color: #CCC;
+	}
+	.pagelink {
+		text-align: center;
+	}
+	.pagelink a {
+		width: 150px;
+	}
 </style>
 
 <div>
@@ -134,4 +148,5 @@
 <script defer src="//play.pokemonshowdown.com/js/battle.js?a7"></script>
 
 <script defer src="js/utils.js?"></script>
+<script defer src="js/replays-battle.js?"></script>
 <script defer src="js/replays.js?"></script>
